delete: Remove peer_channel_encryptor.rs
Functionality moved to:
- peers/handshake/mod.rs
- peers/handshake/states.rs
- peers/conduit.rs

Fuzz testing rewritten here:
- fuzz/src/peer_crypt.rs
